+++ middleware.go
@@ -22,7 +22,8 @@ func recoveryMiddleware(next Handler) Handler {
 	return func(response *Response, r *Request) {
 		defer func() {
 			if err := recover(); err != nil {
-				response.Error(err)
+				response.err = err
+				response.Status(http.StatusInternalServerError)
 			}
 		}()
 
@@ -54,6 +55,35 @@ func parseRequestMiddleware(next Handler) Handler {
 		}
 		request.Data = data
 		next(response, request)
+	}
+}
+
+// corsMiddleware is the middleware handling CORS, using the options set in the router.
+// This middleware is automatically inserted first to the router's list of middleware
+// if the latter has defined CORS Options.
+func corsMiddleware(next Handler) Handler {
+	return func(response *Response, request *Request) {
+		if request.corsOptions == nil {
+			next(response, request)
+			return
+		}
+
+		options := request.corsOptions
+		headers := response.Header()
+		requestHeaders := request.Header()
+
+		options.ConfigureCommon(headers, requestHeaders)
+
+		if request.Method() == http.MethodOptions && requestHeaders.Get("Access-Control-Request-Method") != "" {
+			options.HandlePreflight(headers, requestHeaders)
+			if options.OptionsPassthrough {
+				next(response, request)
+			} else {
+				response.WriteHeader(http.StatusNoContent)
+			}
+		} else {
+			next(response, request)
+		}
 	}
 }